// web/apps/msteams/src/components/organization/organizationManageAssets/__generated__/organizationManageAssets_deskTypes_query.graphql.ts
export type organizationManageAssets_deskTypes_query$data = {
  readonly deskTypes: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly description: string | null | undefined;
        readonly id: string;
        readonly name: string;
      };
    }>;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "organizationManageAssets_deskTypes_query";
};
export type organizationManageAssets_deskTypes_query$key = {
  readonly " $data"?: organizationManageAssets_deskTypes_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"organizationManageAssets_deskTypes_query">;
};