// web/apps/msteams/src/components/organization/organizationManageAssets/__generated__/organizationManageAssets_zones_query.graphql.ts
export type organizationManageAssets_zones_query$data = {
  readonly zones: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly description: string | null | undefined;
        readonly id: string;
        readonly name: string;
      };
    }>;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "organizationManageAssets_zones_query";
};
export type organizationManageAssets_zones_query$key = {
  readonly " $data"?: organizationManageAssets_zones_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"organizationManageAssets_zones_query">;
};